Parse blocks using new structs and disable code we will not be using
